+++ extensions/menubar-weather/src/utils/common-utils.ts
@@ -1,14 +1,17 @@
 import { Cache, getPreferenceValues, Icon } from "@raycast/api";
 import { Preferences } from "../types/preferences";
-import { cityName, latitude, longitude } from "./weather-utils";
+import { cityName, latitude, longitude, showForecast, showLocation, showSun } from "./weather-utils";
 
 export enum CacheKey {
-  CURRENT_WEATHER = "Current Weather",
+  CURRENT_WEATHER = "Open-Meteo Weather",
   LOCATION = "Location",
   REFRESH_TIME = "Refresh Time",
   CITY_NAME = "City Name",
   LONGITUDE = "Longitude",
   LATITUDE = "Latitude",
+  SHOW_SUN = "Show Sun",
+  SHOW_LOCATION = "Show Location",
+  SHOW_FORECAST = "Show Forecast",
 }
 
 export const isEmpty = (string: string | null | undefined) => {
@@ -28,40 +31,6 @@ export function getUnits() {
   return { tempUnit: tempUint, windUint: windUint };
 }
 
-export function getWeatherIcon(icon: string | undefined) {
-  if (typeof icon === "string") {
-    if (icon === "01d" || icon === "01n") {
-      return Icon.Sun;
-    }
-    if (icon === "02d" || icon === "02n") {
-      return Icon.CloudSun;
-    }
-    if (icon === "03d" || icon === "03n") {
-      return Icon.Cloud;
-    }
-    if (icon === "04d" || icon === "04n") {
-      return Icon.Cloud;
-    }
-    if (icon === "09d" || icon === "09n") {
-      return Icon.CloudRain;
-    }
-    if (icon === "10d" || icon === "10n") {
-      return Icon.CloudRain;
-    }
-    if (icon === "11d" || icon === "11n") {
-      return Icon.CloudLightning;
-    }
-    if (icon === "13d" || icon === "13n") {
-      return Icon.Snowflake;
-    }
-    if (icon === "50d" || icon === "50n") {
-      return Icon.Snippets;
-    }
-  } else {
-    return Icon.Sunrise;
-  }
-}
-
 export function isoToDateTime(time: string) {
   return time.replace("T", " ");
 }
@@ -75,15 +44,9 @@ export function timeHour() {
   return date.getHours();
 }
 
-export function getTime(stamp: number) {
-  const timeStamp = stamp * 1000;
-  const date = new Date(timeStamp);
-  return date.toLocaleTimeString();
-}
-
 export function shouldRefresh(oldRefreshTime: number, newRefreshTime: number) {
   const time = newRefreshTime - oldRefreshTime;
-  return time >= 5 * 60 * 1000;
+  return time >= 10 * 60 * 1000;
 }
 
 export function preferencesChanged() {
@@ -93,6 +56,10 @@ export function preferencesChanged() {
   let oldLon = "";
   let oldLat = "";
 
+  let oldShowSun = true;
+  let oldShowLocation = true;
+  let oldShowForecast = true;
+
   const newLon = typeof longitude === "undefined" ? "" : longitude;
   const newLat = typeof latitude === "undefined" ? "" : latitude;
   const newCityName = typeof cityName === "undefined" ? "" : cityName;
@@ -100,6 +67,9 @@ export function preferencesChanged() {
   const cacheCityName = cache.get(CacheKey.CITY_NAME);
   const cacheLon = cache.get(CacheKey.LONGITUDE);
   const cacheLat = cache.get(CacheKey.LATITUDE);
+  const cacheShowSun = cache.get(CacheKey.SHOW_SUN);
+  const cacheShowLocation = cache.get(CacheKey.SHOW_LOCATION);
+  const cacheShowForecast = cache.get(CacheKey.SHOW_FORECAST);
   if (typeof cacheCityName !== "undefined" && !isEmpty(cacheCityName)) {
     oldCityName = JSON.parse(cacheCityName) as string;
   }
@@ -109,10 +79,98 @@ export function preferencesChanged() {
   if (typeof cacheLat !== "undefined" && !isEmpty(cacheLat)) {
     oldLat = JSON.parse(cacheLat) as string;
   }
+  if (typeof cacheShowSun !== "undefined" && !isEmpty(cacheShowSun)) {
+    oldShowSun = JSON.parse(cacheShowSun) as boolean;
+  }
+  if (typeof cacheShowLocation !== "undefined" && !isEmpty(cacheShowLocation)) {
+    oldShowLocation = JSON.parse(cacheShowLocation) as boolean;
+  }
+  if (typeof cacheShowForecast !== "undefined" && !isEmpty(cacheShowForecast)) {
+    oldShowForecast = JSON.parse(cacheShowForecast) as boolean;
+  }
 
   cache.set(CacheKey.CITY_NAME, JSON.stringify(newCityName));
   cache.set(CacheKey.LONGITUDE, JSON.stringify(newLon));
   cache.set(CacheKey.LATITUDE, JSON.stringify(newLat));
+  cache.set(CacheKey.SHOW_SUN, JSON.stringify(showSun));
+  cache.set(CacheKey.SHOW_LOCATION, JSON.stringify(showLocation));
+  cache.set(CacheKey.SHOW_FORECAST, JSON.stringify(showForecast));
 
-  return oldCityName !== newCityName || oldLon !== newLon || oldLat !== newLat;
+  return (
+    oldCityName !== newCityName ||
+    oldLon !== newLon ||
+    oldLat !== newLat ||
+    oldShowSun !== showSun ||
+    oldShowLocation !== showLocation ||
+    oldShowForecast !== showForecast
+  );
+}
+
+export function getDateIcon(day: string) {
+  switch (day) {
+    case "01":
+      return Icon.Number01;
+    case "02":
+      return Icon.Number02;
+    case "03":
+      return Icon.Number03;
+    case "04":
+      return Icon.Number04;
+    case "05":
+      return Icon.Number05;
+    case "06":
+      return Icon.Number06;
+    case "07":
+      return Icon.Number07;
+    case "08":
+      return Icon.Number08;
+    case "09":
+      return Icon.Number09;
+    case "10":
+      return Icon.Number10;
+    case "11":
+      return Icon.Number11;
+    case "12":
+      return Icon.Number12;
+    case "13":
+      return Icon.Number13;
+    case "14":
+      return Icon.Number14;
+    case "15":
+      return Icon.Number15;
+    case "16":
+      return Icon.Number16;
+    case "17":
+      return Icon.Number17;
+    case "18":
+      return Icon.Number18;
+    case "19":
+      return Icon.Number19;
+    case "20":
+      return Icon.Number20;
+    case "21":
+      return Icon.Number21;
+    case "22":
+      return Icon.Number22;
+    case "23":
+      return Icon.Number23;
+    case "24":
+      return Icon.Number24;
+    case "25":
+      return Icon.Number25;
+    case "26":
+      return Icon.Number26;
+    case "27":
+      return Icon.Number27;
+    case "28":
+      return Icon.Number28;
+    case "29":
+      return Icon.Number29;
+    case "30":
+      return Icon.Number30;
+    case "31":
+      return Icon.Number31;
+    default:
+      return Icon.Number00;
+  }
 }