+++ extensions/menubar-weather/src/utils/weather-utils.ts
@@ -4,13 +4,17 @@ import { isEmpty } from "./common-utils";
 import { getOpenMeteoLocation, getOpenMeteoWeather } from "./axios-utils";
 import { GeoLocation } from "../types/types";
 
-export const { cityName, tempUnits, precipitationUnits, longitude, latitude } = getPreferenceValues<Preferences>();
+export const { cityName, tempUnits, precipitationUnits, longitude, latitude, showSun, showLocation, showForecast } =
+  getPreferenceValues<Preferences>();
 
 export async function getGeoLocation() {
   const geoLocation = await getOpenMeteoLocation();
-  const geo = geoLocation.results;
-  if (typeof geoLocation.results !== "undefined" && geoLocation.results.length >= 1) {
-    return geo[0] as GeoLocation;
+  if (
+    typeof geoLocation !== "undefined" &&
+    typeof geoLocation.results !== "undefined" &&
+    geoLocation.results.length >= 1
+  ) {
+    return geoLocation.results[0] as GeoLocation;
   } else
     return {
       id: 5128581,
@@ -65,7 +69,15 @@ export function isEmptyLonLat() {
  */
 export function getWeatherDescription(weatherCode: number | undefined) {
   if (typeof weatherCode === "undefined") {
-    return { description: "No weather info", icon: Icon.Sunrise };
+    return {
+      description: "No weather info",
+      icon: {
+        source: {
+          light: "menubar-icon.png",
+          dark: "menubar-icon@dark.png",
+        },
+      },
+    };
   }
   switch (weatherCode) {
     case 0:
@@ -111,7 +123,7 @@ export function getWeatherDescription(weatherCode: number | undefined) {
     case 73:
       return { description: "Moderate snow fall", icon: Icon.CloudSnow };
     case 75:
-      return { description: "Heavy intensity fall", icon: Icon.CloudSnow };
+      return { description: "Heavy intensity snow fall", icon: Icon.CloudSnow };
     case 77:
       return { description: "Snow grains", icon: Icon.CloudSnow };
     case 80:
@@ -137,6 +149,14 @@ export function getWeatherDescription(weatherCode: number | undefined) {
         icon: Icon.CloudLightning,
       };
     default:
-      return { description: "No weather info", icon: Icon.Sunrise };
+      return {
+        description: "No weather info",
+        icon: {
+          source: {
+            light: "menubar-icon.png",
+            dark: "menubar-icon@dark.png",
+          },
+        },
+      };
   }
 }